+++ Content.Shared/Roles/Actions/MimePowersComponent.cs
@@ -2,7 +2,7 @@ using Robust.Shared.Prototypes;
 using Robust.Shared.Serialization.TypeSerializers.Implementations.Custom;
 using Robust.Shared.Serialization.TypeSerializers.Implementations.Custom.Prototype;
 
-namespace Content.Server.Abilities.Mime;
+namespace Content.Shared.Roles.Actions;
 
 /// <summary>
 /// Lets its owner entity use mime powers, like placing invisible walls.
